[clang-tidy] Generate idedata during ESP-IDF compile
Move the compile_commands.json -> idedata transform into the esphome package
(esphome/espidf/idedata.py) and add espidf.toolchain.get_idedata(), which
parses the build's compile_commands.json (emitted by the IDF cmake build) and
caches it at the same path PlatformIO uses.

Wire it into compile_program's ESP-IDF branch so a normal 'esphome compile'
produces idedata as a build artifact (best-effort) -- restoring IDE
IntelliSense parity for the IDF toolchain (which otherwise emits no idedata)
and giving clang-tidy a real idedata source. script/helpers_idf.py now
delegates to the package so the transform lives in one place.
